First pass at adding OpenType rendering
Later, the classes at the top of the master.feax file can be moved
to an external XML file, so both the OpenType and Graphite code
can use the same classes without duplicating the list of glyphs.

The commented out rules in master.feax are for classes that need to be
defined. Hopefully, the definitions of those classes can be scripted.