Don't use the U AP
This improves the rendering (the two biggest improvements were with
words with the oomatra). It also changes the rendering in many other
test cases, but in smaller amounts that might not be visible.

It should also make the Graphite build give more consistent results.
Before, matras could be positioned with both the U and the M AP,
and depending on the order that make_gdl output the positioning rules,
the rendering would be different.