Use kerning instead of mark positioning for spacing matras
Non-spacing matras still use mark positioning,
and since OpenType zeros any width in a mark that is positioned,
the width of a cluster that has a non-spacing matra might change
slightly. Previously, the RightGuard lookup would add the width
of the matra (spacing or non-spacing) that was to the right of
the anchor in the matra.

Now, for spacing matras, kerning is used to make the anchors line
up, and the regular advance of the matra is added to the cluster
rather than having to be calculated in OpenType.

Non-spacing matras, as the name implies, do not need to add any
space to a cluster.