Review anchors
Most of the changes were one unit differences for consistency.

Changing the anchor from _M to C in iimatra_anusvara
avoided a clash that was always in the OpenType shaping
since it was not in the list of matras. When switching
to kerning the spacing matras, this glyphs was missed
for changing the anchor name.